moves type mappers into destination capabilities
