+++ dlt/common/destination/__init__.py
@@ -2,15 +2,17 @@ from dlt.common.destination.capabilities import (
     DestinationCapabilitiesContext,
     merge_caps_file_formats,
     TLoaderFileFormat,
-    ALL_SUPPORTED_FILE_FORMATS,
+    LOADER_FILE_FORMATS,
 )
 from dlt.common.destination.reference import TDestinationReferenceArg, Destination, TDestination
+from dlt.common.destination.typing import PreparedTableSchema
 
 __all__ = [
     "DestinationCapabilitiesContext",
     "merge_caps_file_formats",
     "TLoaderFileFormat",
-    "ALL_SUPPORTED_FILE_FORMATS",
+    "LOADER_FILE_FORMATS",
+    "PreparedTableSchema",
     "TDestinationReferenceArg",
     "Destination",
     "TDestination",
