+++ dlt/destinations/impl/athena/factory.py
@@ -1,5 +1,6 @@
 import typing as t
 
+from dlt.common.data_types.typing import TDataType
 from dlt.common.destination import Destination, DestinationCapabilitiesContext
 from dlt.common.configuration.specs import AwsCredentials
 from dlt.common.data_writers.escape import (
@@ -7,8 +8,13 @@ from dlt.common.data_writers.escape import (
     format_bigquery_datetime_literal,
 )
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
+from dlt.common.schema.typing import TColumnSchema, TColumnType, TLoaderMergeStrategy, TTableSchema
+from dlt.common.typing import TLoaderFileFormat
+from dlt.common.utils import without_none
 
-from dlt.common.schema.typing import TLoaderMergeStrategy, TTableSchema
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.athena.configuration import AthenaClientConfiguration
 
 if t.TYPE_CHECKING:
@@ -27,6 +33,76 @@ def athena_merge_strategies_selector(
         return []
 
 
+class AthenaTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "complex": "string",
+        "text": "string",
+        "double": "double",
+        "bool": "boolean",
+        "date": "date",
+        "timestamp": "timestamp",
+        "bigint": "bigint",
+        "binary": "binary",
+        "time": "string",
+    }
+
+    sct_to_dbt = {"decimal": "decimal(%i,%i)", "wei": "decimal(%i,%i)"}
+
+    dbt_to_sct = {
+        "varchar": "text",
+        "double": "double",
+        "boolean": "bool",
+        "date": "date",
+        "timestamp": "timestamp",
+        "bigint": "bigint",
+        "binary": "binary",
+        "varbinary": "binary",
+        "decimal": "decimal",
+        "tinyint": "bigint",
+        "smallint": "bigint",
+        "int": "bigint",
+    }
+
+    def ensure_supported_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema,
+        loader_file_format: TLoaderFileFormat,
+    ) -> None:
+        # TIME is not supported for parquet on Athena
+        if loader_file_format == "parquet" and column["data_type"] == "time":
+            raise TerminalValueError(
+                "Please convert `datetime.time` objects in your data to `str` or"
+                " `datetime.datetime`.",
+                "time",
+            )
+
+    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+        precision = column.get("precision")
+        table_format = table.get("table_format")
+        if precision is None:
+            return "bigint"
+        if precision <= 8:
+            return "int" if table_format == "iceberg" else "tinyint"
+        elif precision <= 16:
+            return "int" if table_format == "iceberg" else "smallint"
+        elif precision <= 32:
+            return "int"
+        elif precision <= 64:
+            return "bigint"
+        raise TerminalValueError(
+            f"bigint with {precision} bits precision cannot be mapped into athena integer type"
+        )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int], scale: t.Optional[int]
+    ) -> TColumnType:
+        for key, val in self.dbt_to_sct.items():
+            if db_type.startswith(key):
+                return without_none(dict(data_type=val, precision=precision, scale=scale))  # type: ignore[return-value]
+        return dict(data_type=None)
+
+
 class athena(Destination[AthenaClientConfiguration, "AthenaClient"]):
     spec = AthenaClientConfiguration
 
@@ -37,7 +113,9 @@ class athena(Destination[AthenaClientConfiguration, "AthenaClient"]):
         caps.supported_loader_file_formats = []
         caps.supported_table_formats = ["iceberg", "hive"]
         caps.preferred_staging_file_format = "parquet"
-        caps.supported_staging_file_formats = ["parquet", "jsonl"]
+        caps.supported_staging_file_formats = ["parquet"]
+        caps.type_mapper = AthenaTypeMapper
+
         # athena is storing all identifiers in lower case and is case insensitive
         # it also uses lower case in all the queries
         # https://docs.aws.amazon.com/athena/latest/ug/tables-databases-columns-names.html
