+++ dlt/destinations/impl/bigquery/bigquery.py
@@ -13,25 +13,21 @@ from google.cloud.bigquery.retry import _RETRYABLE_REASONS
 from dlt.common import logger
 from dlt.common.runtime.signals import sleep
 from dlt.common.json import json
-from dlt.common.destination import DestinationCapabilitiesContext
+from dlt.common.destination import DestinationCapabilitiesContext, PreparedTableSchema
 from dlt.common.destination.reference import (
     HasFollowupJobs,
     FollowupJobRequest,
-    PreparedTableSchema,
     TLoadJobState,
     RunnableLoadJob,
     SupportsStagingDestination,
     LoadJob,
 )
 from dlt.common.schema import TColumnSchema, Schema, TTableSchemaColumns
-from dlt.common.schema.typing import TColumnType, TTableFormat
+from dlt.common.schema.typing import TColumnType
 from dlt.common.schema.utils import get_inherited_table_hint
-from dlt.common.schema.utils import table_schema_has_type
-from dlt.common.storages.file_storage import FileStorage
 from dlt.common.storages.load_package import destination_state
 from dlt.common.typing import DictStrAny
 from dlt.destinations.job_impl import DestinationJsonlLoadJob, DestinationParquetLoadJob
-from dlt.destinations.sql_client import SqlClientBase
 from dlt.destinations.exceptions import (
     DatabaseTransientException,
     DatabaseUndefinedRelation,
@@ -48,64 +44,13 @@ from dlt.destinations.impl.bigquery.bigquery_adapter import (
     ROUND_HALF_EVEN_HINT,
     ROUND_HALF_AWAY_FROM_ZERO_HINT,
     TABLE_EXPIRATION_HINT,
+    should_autodetect_schema,
 )
 from dlt.destinations.impl.bigquery.configuration import BigQueryClientConfiguration
 from dlt.destinations.impl.bigquery.sql_client import BigQuerySqlClient, BQ_TERMINAL_REASONS
 from dlt.destinations.job_client_impl import SqlJobClientWithStagingDataset
 from dlt.destinations.job_impl import ReferenceFollowupJobRequest
 from dlt.destinations.sql_jobs import SqlMergeFollowupJob
-from dlt.destinations.type_mapping import TypeMapper
-from dlt.destinations.utils import parse_db_data_type_str_with_precision
-
-
-class BigQueryTypeMapper(TypeMapper):
-    sct_to_unbound_dbt = {
-        "complex": "JSON",
-        "text": "STRING",
-        "double": "FLOAT64",
-        "bool": "BOOL",
-        "date": "DATE",
-        "timestamp": "TIMESTAMP",
-        "bigint": "INT64",
-        "binary": "BYTES",
-        "wei": "BIGNUMERIC",  # non-parametrized should hold wei values
-        "time": "TIME",
-    }
-
-    sct_to_dbt = {
-        "text": "STRING(%i)",
-        "binary": "BYTES(%i)",
-    }
-
-    dbt_to_sct = {
-        "STRING": "text",
-        "FLOAT64": "double",
-        "BOOL": "bool",
-        "DATE": "date",
-        "TIMESTAMP": "timestamp",
-        "INT64": "bigint",
-        "BYTES": "binary",
-        "NUMERIC": "decimal",
-        "BIGNUMERIC": "decimal",
-        "JSON": "complex",
-        "TIME": "time",
-    }
-
-    def to_db_decimal_type(self, column: TColumnSchema) -> str:
-        # Use BigQuery's BIGNUMERIC for large precision decimals
-        precision, scale = self.decimal_precision(column.get("precision"), column.get("scale"))
-        if precision > 38 or scale > 9:
-            return "BIGNUMERIC(%i,%i)" % (precision, scale)
-        return "NUMERIC(%i,%i)" % (precision, scale)
-
-    # noinspection PyTypeChecker,PydanticTypeChecker
-    def from_db_type(
-        self, db_type: str, precision: Optional[int], scale: Optional[int]
-    ) -> TColumnType:
-        # precision is present in the type name
-        if db_type == "BIGNUMERIC":
-            return dict(data_type="wei")
-        return super().from_db_type(*parse_db_data_type_str_with_precision(db_type))
 
 
 class BigQueryLoadJob(RunnableLoadJob, HasFollowupJobs):
@@ -233,7 +178,7 @@ class BigQueryClient(SqlJobClientWithStagingDataset, SupportsStagingDestination)
         super().__init__(schema, config, sql_client)
         self.config: BigQueryClientConfiguration = config
         self.sql_client: BigQuerySqlClient = sql_client  # type: ignore
-        self.type_mapper = BigQueryTypeMapper(self.capabilities)
+        self.type_mapper = self.capabilities.get_type_mapper()
 
     def _create_merge_followup_jobs(
         self, table_chain: Sequence[PreparedTableSchema]
@@ -284,10 +229,10 @@ class BigQueryClient(SqlJobClientWithStagingDataset, SupportsStagingDestination)
     ) -> List[str]:
         # return empty columns which will skip table CREATE or ALTER
         # to let BigQuery autodetect table from data
-        if self._should_autodetect_schema(table_name):
+        table = self.prepare_load_table(table_name)
+        if should_autodetect_schema(table):
             return []
 
-        table = self.prepare_load_table(table_name)
         sql = super()._get_table_update_sql(table_name, new_columns, generate_alter)
         canonical_name = self.sql_client.make_qualified_table_name(table_name)
 
@@ -357,13 +302,21 @@ class BigQueryClient(SqlJobClientWithStagingDataset, SupportsStagingDestination)
 
     def prepare_load_table(self, table_name: str) -> Optional[PreparedTableSchema]:
         table = super().prepare_load_table(table_name)
-        if table_name in self.schema.data_table_names():
+        if table_name not in self.schema.dlt_table_names():
             if TABLE_DESCRIPTION_HINT not in table:
                 table[TABLE_DESCRIPTION_HINT] = (  # type: ignore[name-defined, typeddict-unknown-key, unused-ignore]
                     get_inherited_table_hint(
                         self.schema.tables, table_name, TABLE_DESCRIPTION_HINT, allow_none=True
                     )
                 )
+            if AUTODETECT_SCHEMA_HINT not in table:
+                table[AUTODETECT_SCHEMA_HINT] = (  # type: ignore[typeddict-unknown-key]
+                    get_inherited_table_hint(
+                        self.schema.tables, table_name, AUTODETECT_SCHEMA_HINT, allow_none=True
+                    )
+                    or self.config.autodetect_schema
+                )
+
         return table
 
     def get_storage_tables(
@@ -419,7 +372,7 @@ SELECT {",".join(self._get_storage_table_query_columns())}
     def _get_column_def_sql(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
         name = self.sql_client.escape_column_name(column["name"])
         column_def_sql = (
-            f"{name} {self.type_mapper.to_db_type(column, table)} {self._gen_not_null(column.get('nullable', True))}"
+            f"{name} {self.type_mapper.to_destination_type(column, table)} {self._gen_not_null(column.get('nullable', True))}"
         )
         if column.get(ROUND_HALF_EVEN_HINT, False):
             column_def_sql += " OPTIONS (rounding_mode='ROUND_HALF_EVEN')"
@@ -456,19 +409,12 @@ SELECT {",".join(self._get_storage_table_query_columns())}
             ignore_unknown_values=False,
             max_bad_records=0,
         )
-        if self._should_autodetect_schema(table_name):
+        if should_autodetect_schema(table):
             # allow BigQuery to infer and evolve the schema, note that dlt is not
             # creating such tables at all
             job_config.autodetect = True
             job_config.schema_update_options = bigquery.SchemaUpdateOption.ALLOW_FIELD_ADDITION
             job_config.create_disposition = bigquery.CreateDisposition.CREATE_IF_NEEDED
-        elif ext == "parquet" and table_schema_has_type(table, "complex"):
-            # if table contains complex types, we cannot load with parquet
-            raise LoadJobTerminalException(
-                file_path,
-                "Bigquery cannot load into JSON data type from parquet. Enable autodetect_schema in"
-                " config or via BigQuery adapter or use jsonl format instead.",
-            )
 
         if bucket_path:
             return self.sql_client.native_connection.load_table_from_uri(
@@ -495,12 +441,7 @@ SELECT {",".join(self._get_storage_table_query_columns())}
     def _from_db_type(
         self, bq_t: str, precision: Optional[int], scale: Optional[int]
     ) -> TColumnType:
-        return self.type_mapper.from_db_type(bq_t, precision, scale)
-
-    def _should_autodetect_schema(self, table_name: str) -> bool:
-        return get_inherited_table_hint(
-            self.schema.tables, table_name, AUTODETECT_SCHEMA_HINT, allow_none=True
-        ) or (self.config.autodetect_schema and table_name not in self.schema.dlt_table_names())
+        return self.type_mapper.from_destination_type(bq_t, precision, scale)
 
     def should_truncate_table_before_load_on_staging_destination(self, table_name: str) -> bool:
         return self.config.truncate_tables_on_staging_destination_before_load
