+++ dlt/destinations/impl/clickhouse/clickhouse.py
@@ -29,6 +29,7 @@ from dlt.common.schema.typing import (
     TTableFormat,
     TColumnType,
 )
+from dlt.common.schema.utils import is_nullable_column
 from dlt.common.storages import FileStorage
 from dlt.destinations.exceptions import LoadJobTerminalException
 from dlt.destinations.impl.clickhouse.configuration import (
@@ -54,74 +55,6 @@ from dlt.destinations.job_client_impl import (
 )
 from dlt.destinations.job_impl import ReferenceFollowupJobRequest, FinalizedLoadJobWithFollowupJobs
 from dlt.destinations.sql_jobs import SqlMergeFollowupJob
-from dlt.destinations.type_mapping import TypeMapper
-
-
-class ClickHouseTypeMapper(TypeMapper):
-    sct_to_unbound_dbt = {
-        "complex": "String",
-        "text": "String",
-        "double": "Float64",
-        "bool": "Boolean",
-        "date": "Date",
-        "timestamp": "DateTime64(6,'UTC')",
-        "time": "String",
-        "bigint": "Int64",
-        "binary": "String",
-        "wei": "Decimal",
-    }
-
-    sct_to_dbt = {
-        "decimal": "Decimal(%i,%i)",
-        "wei": "Decimal(%i,%i)",
-        "timestamp": "DateTime64(%i,'UTC')",
-    }
-
-    dbt_to_sct = {
-        "String": "text",
-        "Float64": "double",
-        "Bool": "bool",
-        "Date": "date",
-        "DateTime": "timestamp",
-        "DateTime64": "timestamp",
-        "Time": "timestamp",
-        "Int64": "bigint",
-        "Object('json')": "complex",
-        "Decimal": "decimal",
-    }
-
-    def from_db_type(
-        self, db_type: str, precision: Optional[int] = None, scale: Optional[int] = None
-    ) -> TColumnType:
-        # Remove "Nullable" wrapper.
-        db_type = re.sub(r"^Nullable\((?P<type>.+)\)$", r"\g<type>", db_type)
-
-        # Remove timezone details.
-        if db_type == "DateTime('UTC')":
-            db_type = "DateTime"
-        if datetime_match := re.match(
-            r"DateTime64(?:\((?P<precision>\d+)(?:,?\s*'(?P<timezone>UTC)')?\))?",
-            db_type,
-        ):
-            if datetime_match["precision"]:
-                precision = int(datetime_match["precision"])
-            else:
-                precision = None
-            db_type = "DateTime64"
-
-        # Extract precision and scale, parameters and remove from string.
-        if decimal_match := re.match(
-            r"Decimal\((?P<precision>\d+)\s*(?:,\s*(?P<scale>\d+))?\)", db_type
-        ):
-            precision, scale = decimal_match.groups()  # type: ignore[assignment]
-            precision = int(precision)
-            scale = int(scale) if scale else 0
-            db_type = "Decimal"
-
-        if db_type == "Decimal" and (precision, scale) == self.capabilities.wei_precision:
-            return cast(TColumnType, dict(data_type="wei"))
-
-        return super().from_db_type(db_type, precision, scale)
 
 
 class ClickHouseLoadJob(RunnableLoadJob, HasFollowupJobs):
@@ -286,7 +219,7 @@ class ClickHouseClient(SqlJobClientWithStagingDataset, SupportsStagingDestinatio
         super().__init__(schema, config, self.sql_client)
         self.config: ClickHouseClientConfiguration = config
         self.active_hints = deepcopy(HINT_TO_CLICKHOUSE_ATTR)
-        self.type_mapper = ClickHouseTypeMapper(self.capabilities)
+        self.type_mapper = self.capabilities.get_type_mapper()
 
     def _create_merge_followup_jobs(
         self, table_chain: Sequence[PreparedTableSchema]
@@ -307,9 +240,9 @@ class ClickHouseClient(SqlJobClientWithStagingDataset, SupportsStagingDestinatio
         # Alter table statements only accept `Nullable` modifiers.
         # JSON type isn't nullable in ClickHouse.
         type_with_nullability_modifier = (
-            f"Nullable({self.type_mapper.to_db_type(c,table)})"
-            if c.get("nullable", True)
-            else self.type_mapper.to_db_type(c, table)
+            f"Nullable({self.type_mapper.to_destination_type(c,table)})"
+            if is_nullable_column(c)
+            else self.type_mapper.to_destination_type(c, table)
         )
 
         return (
@@ -371,7 +304,7 @@ class ClickHouseClient(SqlJobClientWithStagingDataset, SupportsStagingDestinatio
     def _from_db_type(
         self, ch_t: str, precision: Optional[int], scale: Optional[int]
     ) -> TColumnType:
-        return self.type_mapper.from_db_type(ch_t, precision, scale)
+        return self.type_mapper.from_destination_type(ch_t, precision, scale)
 
     def should_truncate_table_before_load_on_staging_destination(self, table_name: str) -> bool:
         return self.config.truncate_tables_on_staging_destination_before_load
