+++ dlt/destinations/impl/databricks/factory.py
@@ -1,9 +1,15 @@
 import typing as t
 
+from dlt.common.data_types.typing import TDataType
 from dlt.common.destination import Destination, DestinationCapabilitiesContext
 from dlt.common.data_writers.escape import escape_databricks_identifier, escape_databricks_literal
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
+from dlt.common.schema.typing import TColumnSchema, TColumnType, TTableSchema
+from dlt.common.typing import TLoaderFileFormat
 
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.databricks.configuration import (
     DatabricksCredentials,
     DatabricksClientConfiguration,
@@ -11,6 +17,89 @@ from dlt.destinations.impl.databricks.configuration import (
 
 if t.TYPE_CHECKING:
     from dlt.destinations.impl.databricks.databricks import DatabricksClient
+
+
+class DatabricksTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "complex": "STRING",  # Databricks supports complex types like ARRAY
+        "text": "STRING",
+        "double": "DOUBLE",
+        "bool": "BOOLEAN",
+        "date": "DATE",
+        "timestamp": "TIMESTAMP",  # TIMESTAMP for local timezone
+        "bigint": "BIGINT",
+        "binary": "BINARY",
+        "decimal": "DECIMAL",  # DECIMAL(p,s) format
+        "time": "STRING",
+    }
+
+    dbt_to_sct = {
+        "STRING": "text",
+        "DOUBLE": "double",
+        "BOOLEAN": "bool",
+        "DATE": "date",
+        "TIMESTAMP": "timestamp",
+        "BIGINT": "bigint",
+        "INT": "bigint",
+        "SMALLINT": "bigint",
+        "TINYINT": "bigint",
+        "BINARY": "binary",
+        "DECIMAL": "decimal",
+    }
+
+    sct_to_dbt = {
+        "decimal": "DECIMAL(%i,%i)",
+        "wei": "DECIMAL(%i,%i)",
+    }
+
+    def ensure_supported_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema,
+        loader_file_format: TLoaderFileFormat,
+    ) -> None:
+        if loader_file_format == "jsonl" and column["data_type"] in {
+            "decimal",
+            "wei",
+            "binary",
+            "complex",
+            "date",
+        }:
+            raise TerminalValueError("", column["data_type"])
+
+    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+        precision = column.get("precision")
+        if precision is None:
+            return "BIGINT"
+        if precision <= 8:
+            return "TINYINT"
+        if precision <= 16:
+            return "SMALLINT"
+        if precision <= 32:
+            return "INT"
+        if precision <= 64:
+            return "BIGINT"
+        raise TerminalValueError(
+            f"bigint with {precision} bits precision cannot be mapped into databricks integer type"
+        )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int] = None, scale: t.Optional[int] = None
+    ) -> TColumnType:
+        # precision and scale arguments here are meaningless as they're not included separately in information schema
+        # We use full_data_type from databricks which is either in form "typename" or "typename(precision, scale)"
+        type_parts = db_type.split("(")
+        if len(type_parts) > 1:
+            db_type = type_parts[0]
+            scale_str = type_parts[1].strip(")")
+            precision, scale = [int(val) for val in scale_str.split(",")]
+        else:
+            scale = precision = None
+        db_type = db_type.upper()
+        if db_type == "DECIMAL":
+            if (precision, scale) == self.wei_precision():
+                return dict(data_type="wei", precision=precision, scale=scale)
+        return super().from_destination_type(db_type, precision, scale)
 
 
 class databricks(Destination[DatabricksClientConfiguration, "DatabricksClient"]):
@@ -23,6 +112,7 @@ class databricks(Destination[DatabricksClientConfiguration, "DatabricksClient"])
         caps.preferred_staging_file_format = "parquet"
         caps.supported_staging_file_formats = ["jsonl", "parquet"]
         caps.supported_table_formats = ["delta"]
+        caps.type_mapper = DatabricksTypeMapper
         caps.escape_identifier = escape_databricks_identifier
         # databricks identifiers are case insensitive and stored in lower case
         # https://docs.databricks.com/en/sql/language-manual/sql-ref-identifiers.html
