+++ dlt/destinations/impl/dremio/factory.py
@@ -4,6 +4,11 @@ from dlt.common.destination import Destination, DestinationCapabilitiesContext
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
 from dlt.common.data_writers.escape import escape_dremio_identifier
 
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
+from dlt.common.schema.typing import TColumnSchema, TColumnType
+from dlt.common.typing import TLoaderFileFormat
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.dremio.configuration import (
     DremioCredentials,
     DremioClientConfiguration,
@@ -11,6 +16,68 @@ from dlt.destinations.impl.dremio.configuration import (
 
 if t.TYPE_CHECKING:
     from dlt.destinations.impl.dremio.dremio import DremioClient
+
+
+class DremioTypeMapper(TypeMapperImpl):
+    BIGINT_PRECISION = 19
+    sct_to_unbound_dbt = {
+        "complex": "VARCHAR",
+        "text": "VARCHAR",
+        "double": "DOUBLE",
+        "bool": "BOOLEAN",
+        "date": "DATE",
+        "timestamp": "TIMESTAMP",
+        "bigint": "BIGINT",
+        "binary": "VARBINARY",
+        "time": "TIME",
+    }
+
+    sct_to_dbt = {
+        "decimal": "DECIMAL(%i,%i)",
+        "wei": "DECIMAL(%i,%i)",
+    }
+
+    dbt_to_sct = {
+        "VARCHAR": "text",
+        "DOUBLE": "double",
+        "FLOAT": "double",
+        "BOOLEAN": "bool",
+        "DATE": "date",
+        "TIMESTAMP": "timestamp",
+        "VARBINARY": "binary",
+        "BINARY": "binary",
+        "BINARY VARYING": "binary",
+        "VARIANT": "complex",
+        "TIME": "time",
+        "BIGINT": "bigint",
+        "DECIMAL": "decimal",
+    }
+
+    def ensure_supported_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema,
+        loader_file_format: TLoaderFileFormat,
+    ) -> None:
+        if loader_file_format == "insert_values":
+            return
+        if loader_file_format == "parquet":
+            # binary not supported on parquet if precision is set
+            if column.get("precision"):
+                raise TerminalValueError(
+                    "Dremio cannot load fixed width 'binary' columns from parquet files. Switch to"
+                    " other file format or use binary columns without precision.",
+                    "binary",
+                )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int] = None, scale: t.Optional[int] = None
+    ) -> TColumnType:
+        if db_type == "DECIMAL":
+            if (precision, scale) == self.capabilities.wei_precision:
+                return dict(data_type="wei")
+            return dict(data_type="decimal", precision=precision, scale=scale)
+        return super().from_destination_type(db_type, precision, scale)
 
 
 class dremio(Destination[DremioClientConfiguration, "DremioClient"]):
@@ -23,6 +90,7 @@ class dremio(Destination[DremioClientConfiguration, "DremioClient"]):
         caps.preferred_staging_file_format = "parquet"
         caps.supported_staging_file_formats = ["jsonl", "parquet"]
         caps.escape_identifier = escape_dremio_identifier
+        caps.type_mapper = DremioTypeMapper
         # all identifiers are case insensitive but are stored as is
         # https://docs.dremio.com/current/sonar/data-sources
         caps.has_case_sensitive_identifiers = False
