+++ dlt/destinations/impl/duckdb/factory.py
@@ -4,11 +4,122 @@ from dlt.common.destination import Destination, DestinationCapabilitiesContext
 from dlt.common.data_writers.escape import escape_postgres_identifier, escape_duckdb_literal
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
 
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
+from dlt.common.schema.typing import TColumnSchema, TColumnType
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.duckdb.configuration import DuckDbCredentials, DuckDbClientConfiguration
 
 if t.TYPE_CHECKING:
     from duckdb import DuckDBPyConnection
     from dlt.destinations.impl.duckdb.duck import DuckDbClient
+
+
+class DuckDbTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "complex": "JSON",
+        "text": "VARCHAR",
+        "double": "DOUBLE",
+        "bool": "BOOLEAN",
+        "date": "DATE",
+        # Duck does not allow specifying precision on timestamp with tz
+        "timestamp": "TIMESTAMP WITH TIME ZONE",
+        "bigint": "BIGINT",
+        "binary": "BLOB",
+        "time": "TIME",
+    }
+
+    sct_to_dbt = {
+        # VARCHAR(n) is alias for VARCHAR in duckdb
+        # "text": "VARCHAR(%i)",
+        "decimal": "DECIMAL(%i,%i)",
+        "wei": "DECIMAL(%i,%i)",
+    }
+
+    dbt_to_sct = {
+        "VARCHAR": "text",
+        "JSON": "complex",
+        "DOUBLE": "double",
+        "BOOLEAN": "bool",
+        "DATE": "date",
+        "TIMESTAMP WITH TIME ZONE": "timestamp",
+        "BLOB": "binary",
+        "DECIMAL": "decimal",
+        "TIME": "time",
+        # Int types
+        "TINYINT": "bigint",
+        "SMALLINT": "bigint",
+        "INTEGER": "bigint",
+        "BIGINT": "bigint",
+        "HUGEINT": "bigint",
+        "TIMESTAMP_S": "timestamp",
+        "TIMESTAMP_MS": "timestamp",
+        "TIMESTAMP_NS": "timestamp",
+    }
+
+    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+        precision = column.get("precision")
+        if precision is None:
+            return "BIGINT"
+        # Precision is number of bits
+        if precision <= 8:
+            return "TINYINT"
+        elif precision <= 16:
+            return "SMALLINT"
+        elif precision <= 32:
+            return "INTEGER"
+        elif precision <= 64:
+            return "BIGINT"
+        elif precision <= 128:
+            return "HUGEINT"
+        raise TerminalValueError(
+            f"bigint with {precision} bits precision cannot be mapped into duckdb integer type"
+        )
+
+    def to_db_datetime_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema = None,
+    ) -> str:
+        column_name = column["name"]
+        table_name = table["name"]
+        timezone = column.get("timezone", True)
+        precision = column.get("precision")
+
+        if timezone and precision is not None:
+            raise TerminalValueError(
+                f"DuckDB does not support both timezone and precision for column '{column_name}' in"
+                f" table '{table_name}'. To resolve this issue, either set timezone to False or"
+                " None, or use the default precision."
+            )
+
+        if timezone:
+            # default timestamp mapping for timezone
+            return None
+
+        if precision is None or precision == 6:
+            return "TIMESTAMP"
+        elif precision == 0:
+            return "TIMESTAMP_S"
+        elif precision == 3:
+            return "TIMESTAMP_MS"
+        elif precision == 9:
+            return "TIMESTAMP_NS"
+
+        raise TerminalValueError(
+            f"DuckDB does not support precision '{precision}' for '{column_name}' in table"
+            f" '{table_name}'"
+        )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int], scale: t.Optional[int]
+    ) -> TColumnType:
+        # duckdb provides the types with scale and precision
+        db_type = db_type.split("(")[0].upper()
+        if db_type == "DECIMAL":
+            if precision == 38 and scale == 0:
+                return dict(data_type="wei", precision=precision, scale=scale)
+        return super().from_destination_type(db_type, precision, scale)
 
 
 class duckdb(Destination[DuckDbClientConfiguration, "DuckDbClient"]):
@@ -20,6 +131,7 @@ class duckdb(Destination[DuckDbClientConfiguration, "DuckDbClient"]):
         caps.supported_loader_file_formats = ["insert_values", "parquet", "jsonl"]
         caps.preferred_staging_file_format = None
         caps.supported_staging_file_formats = []
+        caps.type_mapper = DuckDbTypeMapper
         caps.escape_identifier = escape_postgres_identifier
         # all identifiers are case insensitive but are stored as is
         caps.escape_literal = escape_duckdb_literal
