+++ dlt/destinations/impl/mssql/factory.py
@@ -1,14 +1,81 @@
 import typing as t
 
 from dlt.common.destination import Destination, DestinationCapabilitiesContext
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
 from dlt.common.normalizers.naming.naming import NamingConvention
 from dlt.common.data_writers.escape import escape_postgres_identifier, escape_mssql_literal
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
 
+from dlt.common.schema.typing import TColumnSchema, TColumnType
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.mssql.configuration import MsSqlCredentials, MsSqlClientConfiguration
 
 if t.TYPE_CHECKING:
     from dlt.destinations.impl.mssql.mssql import MsSqlJobClient
+
+
+class MsSqlTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "complex": "nvarchar(max)",
+        "text": "nvarchar(max)",
+        "double": "float",
+        "bool": "bit",
+        "bigint": "bigint",
+        "binary": "varbinary(max)",
+        "date": "date",
+        "timestamp": "datetimeoffset",
+        "time": "time",
+    }
+
+    sct_to_dbt = {
+        "complex": "nvarchar(%i)",
+        "text": "nvarchar(%i)",
+        "timestamp": "datetimeoffset(%i)",
+        "binary": "varbinary(%i)",
+        "decimal": "decimal(%i,%i)",
+        "time": "time(%i)",
+        "wei": "decimal(%i,%i)",
+    }
+
+    dbt_to_sct = {
+        "nvarchar": "text",
+        "float": "double",
+        "bit": "bool",
+        "datetimeoffset": "timestamp",
+        "date": "date",
+        "bigint": "bigint",
+        "varbinary": "binary",
+        "decimal": "decimal",
+        "time": "time",
+        "tinyint": "bigint",
+        "smallint": "bigint",
+        "int": "bigint",
+    }
+
+    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+        precision = column.get("precision")
+        if precision is None:
+            return "bigint"
+        if precision <= 8:
+            return "tinyint"
+        if precision <= 16:
+            return "smallint"
+        if precision <= 32:
+            return "int"
+        elif precision <= 64:
+            return "bigint"
+        raise TerminalValueError(
+            f"bigint with {precision} bits precision cannot be mapped into mssql integer type"
+        )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int], scale: t.Optional[int]
+    ) -> TColumnType:
+        if db_type == "decimal":
+            if (precision, scale) == self.capabilities.wei_precision:
+                return dict(data_type="wei")
+        return super().from_destination_type(db_type, precision, scale)
 
 
 class mssql(Destination[MsSqlClientConfiguration, "MsSqlJobClient"]):
@@ -20,6 +87,7 @@ class mssql(Destination[MsSqlClientConfiguration, "MsSqlJobClient"]):
         caps.supported_loader_file_formats = ["insert_values"]
         caps.preferred_staging_file_format = None
         caps.supported_staging_file_formats = []
+        caps.type_mapper = MsSqlTypeMapper
         # mssql is by default case insensitive and stores identifiers as is
         # case sensitivity can be changed by database collation so we allow to reconfigure
         # capabilities in the mssql factory
