+++ dlt/destinations/impl/postgres/postgres.py
@@ -18,6 +18,7 @@ from dlt.common.destination import DestinationCapabilitiesContext
 from dlt.common.exceptions import TerminalValueError
 from dlt.common.schema import TColumnSchema, TColumnHint, Schema
 from dlt.common.schema.typing import TColumnType, TTableFormat
+from dlt.common.schema.utils import is_nullable_column
 from dlt.common.storages.file_storage import FileStorage
 
 from dlt.destinations.sql_jobs import SqlStagingCopyFollowupJob, SqlJobParams
@@ -25,103 +26,8 @@ from dlt.destinations.insert_job_client import InsertValuesJobClient
 from dlt.destinations.impl.postgres.sql_client import Psycopg2SqlClient
 from dlt.destinations.impl.postgres.configuration import PostgresClientConfiguration
 from dlt.destinations.sql_client import SqlClientBase
-from dlt.destinations.type_mapping import TypeMapper
 
 HINT_TO_POSTGRES_ATTR: Dict[TColumnHint, str] = {"unique": "UNIQUE"}
-
-
-class PostgresTypeMapper(TypeMapper):
-    sct_to_unbound_dbt = {
-        "complex": "jsonb",
-        "text": "varchar",
-        "double": "double precision",
-        "bool": "boolean",
-        "date": "date",
-        "bigint": "bigint",
-        "binary": "bytea",
-        "timestamp": "timestamp with time zone",
-        "time": "time without time zone",
-    }
-
-    sct_to_dbt = {
-        "text": "varchar(%i)",
-        "timestamp": "timestamp (%i) with time zone",
-        "decimal": "numeric(%i,%i)",
-        "time": "time (%i) without time zone",
-        "wei": "numeric(%i,%i)",
-    }
-
-    dbt_to_sct = {
-        "varchar": "text",
-        "jsonb": "complex",
-        "double precision": "double",
-        "boolean": "bool",
-        "timestamp with time zone": "timestamp",
-        "date": "date",
-        "bigint": "bigint",
-        "bytea": "binary",
-        "numeric": "decimal",
-        "time without time zone": "time",
-        "character varying": "text",
-        "smallint": "bigint",
-        "integer": "bigint",
-    }
-
-    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
-        precision = column.get("precision")
-        if precision is None:
-            return "bigint"
-        # Precision is number of bits
-        if precision <= 16:
-            return "smallint"
-        elif precision <= 32:
-            return "integer"
-        elif precision <= 64:
-            return "bigint"
-        raise TerminalValueError(
-            f"bigint with {precision} bits precision cannot be mapped into postgres integer type"
-        )
-
-    def to_db_datetime_type(
-        self,
-        column: TColumnSchema,
-        table: PreparedTableSchema = None,
-    ) -> str:
-        column_name = column.get("name")
-        table_name = table.get("name")
-        timezone = column.get("timezone")
-        precision = column.get("precision")
-
-        if timezone is None and precision is None:
-            return None
-
-        timestamp = "timestamp"
-
-        # append precision if specified and valid
-        if precision is not None:
-            if 0 <= precision <= 6:
-                timestamp += f" ({precision})"
-            else:
-                raise TerminalValueError(
-                    f"Postgres does not support precision '{precision}' for '{column_name}' in"
-                    f" table '{table_name}'"
-                )
-
-        # append timezone part
-        if timezone is None or timezone:  # timezone True and None
-            timestamp += " with time zone"
-        else:  # timezone is explicitly False
-            timestamp += " without time zone"
-
-        return timestamp
-
-    def from_db_type(
-        self, db_type: str, precision: Optional[int] = None, scale: Optional[int] = None
-    ) -> TColumnType:
-        if db_type == "numeric":
-            if (precision, scale) == self.capabilities.wei_precision:
-                return dict(data_type="wei")
-        return super().from_db_type(db_type, precision, scale)
 
 
 class PostgresStagingCopyJob(SqlStagingCopyFollowupJob):
@@ -203,7 +109,7 @@ class PostgresCsvCopyJob(RunnableLoadJob, HasFollowupJobs):
             for col in self._job_client.schema.get_table_columns(table_name).values():
                 norm_col = sql_client.escape_column_name(col["name"], escape=True)
                 split_columns.append(norm_col)
-                if norm_col in split_headers and col.get("nullable", True):
+                if norm_col in split_headers and is_nullable_column(col):
                     split_null_headers.append(norm_col)
             split_unknown_headers = set(split_headers).difference(split_columns)
             if split_unknown_headers:
@@ -256,7 +162,7 @@ class PostgresClient(InsertValuesJobClient):
         self.config: PostgresClientConfiguration = config
         self.sql_client: Psycopg2SqlClient = sql_client
         self.active_hints = HINT_TO_POSTGRES_ATTR if self.config.create_indexes else {}
-        self.type_mapper = PostgresTypeMapper(self.capabilities)
+        self.type_mapper = self.capabilities.get_type_mapper()
 
     def create_load_job(
         self, table: PreparedTableSchema, file_path: str, load_id: str, restore: bool = False
@@ -274,7 +180,7 @@ class PostgresClient(InsertValuesJobClient):
         )
         column_name = self.sql_client.escape_column_name(c["name"])
         return (
-            f"{column_name} {self.type_mapper.to_db_type(c,table)} {hints_str} {self._gen_not_null(c.get('nullable', True))}"
+            f"{column_name} {self.type_mapper.to_destination_type(c,table)} {hints_str} {self._gen_not_null(c.get('nullable', True))}"
         )
 
     def _create_replace_followup_jobs(
@@ -287,4 +193,4 @@ class PostgresClient(InsertValuesJobClient):
     def _from_db_type(
         self, pq_t: str, precision: Optional[int], scale: Optional[int]
     ) -> TColumnType:
-        return self.type_mapper.from_db_type(pq_t, precision, scale)
+        return self.type_mapper.from_destination_type(pq_t, precision, scale)
