+++ dlt/destinations/impl/redshift/factory.py
@@ -1,10 +1,17 @@
 import typing as t
 
+from dlt.common.data_types.typing import TDataType
 from dlt.common.destination import Destination, DestinationCapabilitiesContext
 from dlt.common.data_writers.escape import escape_redshift_identifier, escape_redshift_literal
 from dlt.common.arithmetics import DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE
 
+from dlt.common.destination.typing import PreparedTableSchema
+from dlt.common.exceptions import TerminalValueError
 from dlt.common.normalizers.naming import NamingConvention
+from dlt.common.schema.typing import TColumnSchema, TColumnType
+from dlt.common.typing import TLoaderFileFormat
+
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.redshift.configuration import (
     RedshiftCredentials,
     RedshiftClientConfiguration,
@@ -12,6 +19,92 @@ from dlt.destinations.impl.redshift.configuration import (
 
 if t.TYPE_CHECKING:
     from dlt.destinations.impl.redshift.redshift import RedshiftClient
+
+
+class RedshiftTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "complex": "super",
+        "text": "varchar(max)",
+        "double": "double precision",
+        "bool": "boolean",
+        "date": "date",
+        "timestamp": "timestamp with time zone",
+        "bigint": "bigint",
+        "binary": "varbinary",
+        "time": "time without time zone",
+    }
+
+    sct_to_dbt = {
+        "decimal": "numeric(%i,%i)",
+        "wei": "numeric(%i,%i)",
+        "text": "varchar(%i)",
+        "binary": "varbinary(%i)",
+    }
+
+    dbt_to_sct = {
+        "super": "complex",
+        "varchar(max)": "text",
+        "double precision": "double",
+        "boolean": "bool",
+        "date": "date",
+        "timestamp with time zone": "timestamp",
+        "bigint": "bigint",
+        "binary varying": "binary",
+        "numeric": "decimal",
+        "time without time zone": "time",
+        "varchar": "text",
+        "smallint": "bigint",
+        "integer": "bigint",
+    }
+
+    def ensure_supported_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema,
+        loader_file_format: TLoaderFileFormat,
+    ) -> None:
+        if loader_file_format == "insert_values":
+            return
+        # time not supported on staging file formats
+        if column["data_type"] == "time":
+            raise TerminalValueError(
+                "Please convert `datetime.time` objects in your data to `str` or"
+                " `datetime.datetime`.",
+                "time",
+            )
+        if loader_file_format == "jsonl":
+            if column["data_type"] == "binary":
+                raise TerminalValueError("", "binary")
+        if loader_file_format == "parquet":
+            # binary not supported on parquet if precision is set
+            if column.get("precision"):
+                raise TerminalValueError(
+                    "Redshift cannot load fixed width VARBYTE columns from parquet files. Switch"
+                    " to other file format or use binary columns without precision.",
+                    "binary",
+                )
+
+    def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+        precision = column.get("precision")
+        if precision is None:
+            return "bigint"
+        if precision <= 16:
+            return "smallint"
+        elif precision <= 32:
+            return "integer"
+        elif precision <= 64:
+            return "bigint"
+        raise TerminalValueError(
+            f"bigint with {precision} bits precision cannot be mapped into postgres integer type"
+        )
+
+    def from_destination_type(
+        self, db_type: str, precision: t.Optional[int], scale: t.Optional[int]
+    ) -> TColumnType:
+        if db_type == "numeric":
+            if (precision, scale) == self.capabilities.wei_precision:
+                return dict(data_type="wei")
+        return super().from_destination_type(db_type, precision, scale)
 
 
 class redshift(Destination[RedshiftClientConfiguration, "RedshiftClient"]):
@@ -23,6 +116,7 @@ class redshift(Destination[RedshiftClientConfiguration, "RedshiftClient"]):
         caps.supported_loader_file_formats = ["insert_values"]
         caps.preferred_staging_file_format = "jsonl"
         caps.supported_staging_file_formats = ["jsonl", "parquet"]
+        caps.type_mapper = RedshiftTypeMapper
         # redshift is case insensitive and will lower case identifiers when stored
         # you can enable case sensitivity https://docs.aws.amazon.com/redshift/latest/dg/r_enable_case_sensitive_identifier.html
         # then redshift behaves like postgres
