+++ dlt/destinations/impl/weaviate/factory.py
@@ -2,6 +2,7 @@ import typing as t
 
 from dlt.common.destination import Destination, DestinationCapabilitiesContext
 
+from dlt.destinations.type_mapping import TypeMapperImpl
 from dlt.destinations.impl.weaviate.configuration import (
     WeaviateCredentials,
     WeaviateClientConfiguration,
@@ -11,6 +12,33 @@ if t.TYPE_CHECKING:
     from dlt.destinations.impl.weaviate.weaviate_client import WeaviateClient
 
 
+class WeaviateTypeMapper(TypeMapperImpl):
+    sct_to_unbound_dbt = {
+        "text": "text",
+        "double": "number",
+        "bool": "boolean",
+        "timestamp": "date",
+        "date": "date",
+        "time": "text",
+        "bigint": "int",
+        "binary": "blob",
+        "decimal": "text",
+        "wei": "number",
+        "complex": "text",
+    }
+
+    sct_to_dbt = {}
+
+    dbt_to_sct = {
+        "text": "text",
+        "number": "double",
+        "boolean": "bool",
+        "date": "timestamp",
+        "int": "bigint",
+        "blob": "binary",
+    }
+
+
 class weaviate(Destination[WeaviateClientConfiguration, "WeaviateClient"]):
     spec = WeaviateClientConfiguration
 
@@ -18,6 +46,7 @@ class weaviate(Destination[WeaviateClientConfiguration, "WeaviateClient"]):
         caps = DestinationCapabilitiesContext()
         caps.preferred_loader_file_format = "jsonl"
         caps.supported_loader_file_formats = ["jsonl"]
+        caps.type_mapper = WeaviateTypeMapper
         # weaviate names are case sensitive following GraphQL naming convention
         # https://weaviate.io/developers/weaviate/config-refs/schema
         caps.has_case_sensitive_identifiers = False
