+++ dlt/destinations/type_mapping.py
@@ -7,13 +7,12 @@ from dlt.common.schema.typing import (
     TDataType,
     TColumnType,
 )
-from dlt.common.destination.capabilities import DestinationCapabilitiesContext
+from dlt.common.destination.capabilities import DataTypeMapper
+from dlt.common.typing import TLoaderFileFormat
 from dlt.common.utils import without_none
 
 
-class TypeMapper:
-    capabilities: DestinationCapabilitiesContext
-
+class TypeMapperImpl(DataTypeMapper):
     sct_to_unbound_dbt: Dict[TDataType, str]
     """Data types without precision or scale specified (e.g. `"text": "varchar"` in postgres)"""
     sct_to_dbt: Dict[TDataType, str]
@@ -23,8 +22,13 @@ class TypeMapper:
 
     dbt_to_sct: Dict[str, TDataType]
 
-    def __init__(self, capabilities: DestinationCapabilitiesContext) -> None:
-        self.capabilities = capabilities
+    def ensure_supported_type(
+        self,
+        column: TColumnSchema,
+        table: PreparedTableSchema,
+        loader_file_format: TLoaderFileFormat,
+    ) -> None:
+        pass
 
     def to_db_integer_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
         # Override in subclass if db supports other integer types (e.g. smallint, integer, tinyint, etc.)
@@ -64,7 +68,7 @@ class TypeMapper:
         return self.sct_to_dbt["decimal"] % (precision_tup[0], precision_tup[1])
 
     # TODO: refactor lancedb and weaviate to make table object required
-    def to_db_type(self, column: TColumnSchema, table: PreparedTableSchema = None) -> str:
+    def to_destination_type(self, column: TColumnSchema, table: PreparedTableSchema) -> str:
         sc_t = column["data_type"]
         if sc_t == "bigint":
             db_t = self.to_db_integer_type(column, table)
@@ -130,7 +134,7 @@ class TypeMapper:
             scale if scale is not None else default_scale,
         )
 
-    def from_db_type(
+    def from_destination_type(
         self, db_type: str, precision: Optional[int], scale: Optional[int]
     ) -> TColumnType:
         return without_none(
